+++ service/singleton/ddns/dummy.go
@@ -1,0 +1,7 @@
+package ddns
+
+type ProviderDummy struct{}
+
+func (provider ProviderDummy) UpdateDomain(domainConfig *DomainConfig) bool {
+	return false
+}
